+++ rpxy-bin/src/config/mod.rs
@@ -3,7 +3,10 @@ mod service;
 mod toml;
 
 pub use {
-  self::toml::ConfigToml,
   parse::{build_cert_manager, build_settings, parse_opts},
   service::ConfigTomlReloader,
+  toml::ConfigToml,
 };
+
+#[cfg(feature = "acme")]
+pub use parse::build_acme_manager;